Method moveInDir works. This allows the birds to fly in a certain angle.
calcAngleToPos works. This calculates the angle necessary to move from the birds position to the set position.

averagePosition works, this calculates the average position of the birds in range and sets the direction towards this average position.

This is now a very crude implementation of birds grouping up when near eachother.